Move EventTimetable exporter outside
