+++ app/controllers/__init__.py
@@ -4,6 +4,8 @@ from .daily_plan_tasks import DailyPlanTaskView
 from .daily_plan_recipes import DailyPlanRecipeView
 from .event_exporter import EventExporterView
 from .exports.event_cookbook_exporter import EventCookbookExporterView
+from .exports.event_timetable_exporter import EventTimetableExporterView
+
 
 from .events.events import EventView
 from .events.share_events import ShareEventView
@@ -44,6 +46,7 @@ __all__ = [
     "EventExporterView",
     "EventView",
     "EventCookbookExporterView",
+    "EventTimetableExporterView",
     "ShareEventView",
     "PublishedEventView",
     "ErrorView",
@@ -84,6 +87,7 @@ def register_all_controllers(application):
 
     EventExporterView.register(application, base_class=HelperFlaskView)
     EventCookbookExporterView.register(application, base_class=HelperFlaskView)
+    EventTimetableExporterView.register(application, base_class=HelperFlaskView)
 
     ErrorView.register(application)
     SupportView.register(application)
