+++ app/templates/events/_export.html.j2
@@ -10,7 +10,7 @@
 			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:show_ingredient_list', event_id=event.id) }}"> {{ icon("list") }} nákupní seznam s recepty </a>
 		</div>
 		<div class="col-12 col-sm-3">
-			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:show_recipe_list', event_id=event.id) }}"> {{ icon("food") }} jídelníček </a>
+			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventTimetableExporterView:show', event_id=event.id) }}"> {{ icon("food") }} jídelníček </a>
 		</div>
 		<div class="col-12 col-sm-3">
 			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventCookbookExporterView:show', event_id=event.id) }}"> {{ icon("cookbook") }} kuchařka </a>
@@ -29,7 +29,7 @@
 			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:show_ingredient_list_pdf', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} nákupní seznam s recepty  </a>
 		</div>
 		<div class="col-12 col-sm-3">
-			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:show_recipe_list_pdf', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} jídelníček  </a>
+			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventTimetableExporterView:show', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} jídelníček  </a>
 		</div>
 		<div class="col-12 col-sm-3">
 			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventCookbookExporterView:pdf', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} kuchařka </a>
@@ -48,7 +48,7 @@
 			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:download_ingredient_list_pdf', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} nákupní seznam s recepty </a>
 		</div>
 		<div class="col-12 col-sm-3">
-			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:download_recipe_list_pdf', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} jídelníček  </a>
+			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventTimetableExporterView:download', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} jídelníček  </a>
 		</div>
 		<div class="col-12 col-sm-3">
 			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventCookbookExporterView:download', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} kuchařka </a>
